Rework networking to use a tap
On my machine, the macvlan does not seem to pick up packages from
other macvlans and I fail to trace down why. An alternative is to
do what a macvlan does manually using a tap device. The added
advantage is that this significantly simplifies the code base, at
the cost of having to use more CPU for sending network packages.